requirements.txt
streamlit
requests
Pillow
pyttsx3
googletrans==4.0.0-rc1
gTTS
